begin connection management revamp
